code/requirements.txt
pyinstaller==1.3.3
pyproj==2.6.1.post1
geopandas==0.8.1
GDAL==3.1.2
Fiona==1.8.13
requests==2.24.0
Shapely==1.7.0
